WIP: Add intro and first step docs
Having clarified the two most common ways of using the connector, three new files were created to take the place of gettingstarted.md. 1. start-here gives a brief overview answering the question "What is the Spark-Riak connector?" 2. download-and-install gives instructions for a user who wants to dl, build, & install locally. 3. add-as-dependency is where instructions for a user who wants to add the connector as a dpenency will go.